+++ backend/data/src/main/kotlin/io/tolgee/component/SentryBeforeSendCallback.kt
@@ -18,6 +18,7 @@ class SentryBeforeSendCallback : SentryOptions.BeforeSendCallback {
       listOf(
         "FailedDontRequeueException",
         "ClientAbortException",
+        "AsyncRequestNotUsableException",
       )
   }